requirements.txt
streamlit==1.33.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
plotly==5.21.0
tensorflow==2.18.0
Pillow==10.2.0
PyPDF2==3.0.1
python-dotenv==1.0.1
google-generativeai==0.4.1
seaborn==0.11.2  
